+++ docs/trace/extending-the-sdk/MyExporterExtensions.cs
@@ -1,4 +1,4 @@
-// <copyright file="MyExporterHelperExtensions.cs" company="OpenTelemetry Authors">
+// <copyright file="MyExporterExtensions.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,7 +18,7 @@ using System;
 using OpenTelemetry;
 using OpenTelemetry.Trace;
 
-internal static class MyExporterHelperExtensions
+internal static class MyExporterExtensions
 {
     public static TracerProviderBuilder AddMyExporter(this TracerProviderBuilder builder)
     {
